requirements.txt
streamlit
pymongo
transformers
sentence-transformers
faiss-cpu
requests
wikipedia-api
numpy
joblib